add docs for custom dataset, reader and model technical
